Added simple JavaScript calculator.
HTML | Calculator
Here a Calculator is going to be formed with HTML code.

Calculator Title: This is the title at the top of our application, “HTML Calculator”.
Output Screen: This will be our output screen, where all text will be shown. Like the input that the user will type and the answer calculated from the user input. So, we can again break down this into two smaller pieces as shown below:
Question Output: This will be the input given by the user.
Answer Output: This will be the result calculated from user input.